\name{arithmatic.fd}
\alias{==.basisfd}
\title{
  Arithmatic on functional objects 
}
\description{
  Arithmatic on functional basis or data objects 
}
\usage{
"==.basisfd"(basis1, basis2)
}
\arguments{
  \item{basis1, basis2}{
    functional basis object 
  }
}

\value{
  basisobj1 == basisobj2 returns a logical scalar.  

}
\details{

}
\seealso{
\code{\link{basisfd}}, 
\code{\link{basisfd.product}}
}
% docclass is function
\keyword{smooth}

