\name{arithmetic.basisfd}
\alias{arithmetic.basisfd}
\alias{==.basisfd}
\title{
  Arithmatic on functional basis objects 
}
\description{
  Arithmatic on functional basis objects 
}
\usage{
"==.basisfd"(basis1, basis2)
}
\arguments{
  \item{basis1, basis2}{
    functional basis object 
  }
}

\value{
  basisobj1 == basisobj2 returns a logical scalar.
}
\details{

}
\seealso{
\code{\link{basisfd}}, 
\code{\link{basisfd.product}}
\code{\link{arithmetic.fd}}
}
% docclass is function
\keyword{smooth}

