\name{center.fd}
\alias{center.fd}
\title{Center Functional Data}
\description{
Subtract the pointwise mean from each of the functions in a functional data object; that is, to center them on the mean function. 
}
\usage{
center.fd(fd)
}
\arguments{
  \item{fd}{A functional data object. }
}
\details{
}
\value{
Returns a functional data object with same dimensions as fd that contains the centered versions of the functions in the object fd. 
}
\references{}
\note{}
\author{}
\seealso{
sumFd, meanFd, std.fd 
}
\examples{}
\keyword{}
