\name{fda-internal}
\title{FDA internal functions}
\description{Internal undocumentation functions}
\usage{center.fd(fdobj)}
\alias{/.fd}
\alias{Varfnden}
\alias{Varfninten}
\alias{[.fd}
\alias{^.fd}
\alias{c.fd}
\alias{derivchk}
\alias{derivs}
\alias{eigchk}
\alias{expect.phiphit}
\alias{fd2list}
\alias{fdchk}
\alias{fngrad.smooth.monotone}
\alias{fngrad.smooth.morph}
\alias{fRegressArgCheck}
\alias{geigen}
\alias{hesscal.smooth.monotone}
\alias{hesscal.smooth.morph}
\alias{is.diag}
\alias{is.eqbasis}
\alias{is.integerLfd}
\alias{isotone}
\alias{knotmultchk}
\alias{linesearch}
\alias{linesearch.smooth.monotone}
\alias{linesearch.smooth.morph}
\alias{loglfnden}
\alias{loglfninten}
\alias{loglfnpos}
\alias{loglhesspos}
\alias{m2ij}
\alias{mongrad}
\alias{monhess}
\alias{normalize.phi}
\alias{pendiagfn}
\alias{plot.cca.fd}
\alias{polintarray}
\alias{polintmat}
\alias{polynom}
\alias{polynompen}
\alias{polyprod}
\alias{ppBspline}
\alias{ppderiv}
\alias{print.Lfd}
\alias{print.basisfd}
\alias{print.bifd}
\alias{print.fd}
\alias{print.fdPar}
\alias{rangechk}
\alias{regfngrad}
\alias{reghess}
\alias{regyfn}
\alias{rkck}
\alias{rkqs}
\alias{shifty}
\alias{sqrt.fd}
\alias{stepchk}
\alias{stepit}
\alias{symsolve}
\alias{trapzmat}
\alias{use.proper.basis}
\alias{PENSSEfun}
\alias{PENSSEhess}
\alias{argcheck}
\alias{fdParcheck}
\alias{trapz}
\alias{ycheck}
\alias{moreNames}

% docclass is function
\keyword{internal}
