\name{fda-interna}
\title{FDA internal functions}
\description{Internal undocumentation functions}
\usage{
center.fd(fdobj)
}
\alias{*.fd}
\alias{+.fd}
\alias{-.fd}
\alias{/.fd}
\alias{Varfnden}
\alias{Varfninten}
\alias{[.fd}
\alias{^.fd}
\alias{bifd}
\alias{c.fd}
\alias{daily}
\alias{derivchk}
\alias{derivs}
\alias{eigchk}
\alias{expect.phi}
\alias{expect.phiphit}
\alias{fd2list}
\alias{fdchk}
\alias{fngrad.smooth.monotone}
\alias{fngrad.smooth.morph}
\alias{gait}
\alias{geigen}
\alias{growth}
\alias{handwrit}
\alias{hesscal.smooth.monotone}
\alias{hesscal.smooth.morph}
\alias{is.diag}
\alias{is.eqbasis}
\alias{is.integer.Lfd}
\alias{isotone}
\alias{knotmultchk}
\alias{linesearch}
\alias{linesearch.smooth.monotone}
\alias{linesearch.smooth.morph}
\alias{lip}
\alias{loglfnden}
\alias{loglfninten}
\alias{loglfnpos}
\alias{loglhesspos}
\alias{m2ij}
\alias{melanoma}
\alias{monfn}
\alias{mongrad}
\alias{monhess}
\alias{nondurables}
\alias{normalize.phi}
\alias{pendiagfn}
\alias{pinch}
\alias{plot.Lfd}
\alias{plot.cca.fd}
\alias{plot.pca.fd}
\alias{polintarray}
\alias{polintmat}
\alias{polynom}
\alias{polynompen}
\alias{polyprod}
\alias{ppBspline}
\alias{ppderiv}
\alias{print.Lfd}
\alias{print.basisfd}
\alias{print.bifd}
\alias{print.fd}
\alias{print.fdPar}
\alias{rangechk}
\alias{refinery}
\alias{regfngrad}
\alias{reghess}
\alias{regyfn}
\alias{rkck}
\alias{rkqs}
\alias{shifty}
\alias{sqrt.fd}
\alias{stepchk}
\alias{stepit}
\alias{symsolve}
\alias{trapzmat}
\alias{use.proper.basis}
\alias{zerofind}

% docclass is function
\keyword{internal}
