\name{fda.internal}
\alias{c.fd}
\alias{create.basis.fd}
\alias{create.default.basis} 
\alias{create.polynomial.basis}   
\alias{cycleplot.fd}  
\alias{derivchk}                  
\alias{derivs} 
\alias{expect.phi}                
\alias{expect.phiphit}  
\alias{^.fd}    
\alias{-.fd}     
\alias{/.fd}  
\alias{[.fd}
\alias{*.fd} 
\alias{+.fd}
\alias{fd2data}                   
\alias{fngrad.smooth.monotone}     
\alias{geigen}                    
\alias{getnames} 
\alias{hesscal.smooth.monotone}   
\alias{inprod.n}   
\alias{interpreg}                 
\alias{is.diag}     
\alias{is.fd}                     
\alias{is.Lfd} 
\alias{isotone}                   
\alias{linesearch}                 
\alias{linesearch.smooth.monotone}
\alias{linesSmoothPspline}       
\alias{loglfnden}                 
\alias{loglfnpos}                  
\alias{m2ij}                      
\alias{make.basis}                 
\alias{monfn}                     
\alias{monfneval}            
\alias{mongrad}                   
\alias{mongrad1}      
\alias{monhess}                   
\alias{normalize.phi}          
\alias{odesolv}                   
\alias{pda.fd}    
\alias{pdalist}                   
\alias{pendiagfn}    
\alias{plotscores}                
\alias{plotSmoothPspline}   
\alias{polintarray}               
\alias{polintmat}    
\alias{polynom}  
\alias{polynompen}             
\alias{predictSmoothPspline}    
\alias{print.smooth.Pspline}    
\alias{ptwiseLM}                  
\alias{putcoef}    
\alias{rangechk}    
\alias{regfngrad}                  
\alias{reghess}                   
\alias{register}                   
\alias{regyfn}                    
\alias{rkck}                       
\alias{rkqs}                      
\alias{shifty}                     
\alias{smooth.basis.n}            
\alias{smooth.pos}                
\alias{sqrt.fd}                   
\alias{SRsetup}                    
\alias{SRsetup1}                  
\alias{stepchk}                    
\alias{stepit}                    
\alias{symsolve}                   
\alias{trapzmat}                  
\alias{trapz.n}                    
\alias{use.proper.basis}          
\alias{Varfnden}                
\alias{Varfnpos}                  
\alias{varmx}                      
\alias{warpfngrad}                
\alias{warphesscal}               
\alias{zerofind}                  
\title{Internal Functions and Objects}
\description{
These are internal functions and objects of the FDA package and should never be used by users in any circumstance. 
}
\usage{
fda.internal
}
\arguments{}
\details{}
\value{}
\references{}
\note{}
\author{}
\seealso{}
\examples{}
\keyword{internal}
