\name{getbasis}
\alias{getbasis}
\title{Get the Basis Object from a Functional Data Object }
\description{
Retrieves the basis defining a functional data object. 
}
\usage{
getbasis(fd)
}
\arguments{
  \item{fd }{A functional data object. }
}
\details{
}
\value{
The basis object defining fd. 
}
\references{}
\note{}
\author{}
\seealso{}
\examples{
}
\keyword{}
