\name{getbasisrange}
\alias{getbasisrange}
\title{
  Extract the range from a basis object 
}
\description{
  Extracts the 'range' component from basis object 'basisobj'.  
}
\usage{
getbasisrange(basisobj) 
}
\arguments{
  \item{basisobj}{
    a functional basis object 
  }
}

\value{
  a numeric vector of length 2 
}
\references{
  Ramsay, James O., Hooker, Giles, and Graves, Spencer (2009),
    \emph{Functional data analysis with R and Matlab}, Springer, New York.
  
  Ramsay, James O., and Silverman, Bernard W. (2005), 
    \emph{Functional Data Analysis, 2nd ed.}, Springer, New York.
  
  Ramsay, James O., and Silverman, Bernard W. (2002), 
    \emph{Applied Functional Data Analysis}, Springer, New York.
}
\keyword{attribute}
