\name{getbasistype}
\alias{getbasistype}
\title{Get the Basis Type from a Basis Object }
\description{
Retrieves the basis type from a basis object. 
}
\usage{
getbasistype(basisfd)
}
\arguments{
  \item{basisfd}{A basis object.}
}
\details{
}
\value{
A string such as "bspline" or "fourier" specifying the basis type. 
}
\references{}
\note{}
\author{}
\seealso{}
\examples{
}
\keyword{}
