\name{getcoef}
\alias{getcoef}
\title{Get the Coefficient Array from a Functional Data Object }
\description{
Retrieves the coefficient array defining a functional data object. 
}
\usage{
getcoef(fd)
}
\arguments{
  \item{fd }{A functional data object. }
}
\details{
}
\value{
A 1-, 2-, or 3-dimensional array of coefficients defining a functional data object. 
}
\references{}
\note{}
\author{}
\seealso{}
\examples{
}
\keyword{}
