\name{is.fd}
\alias{is.fd}
\title{
  Confirm Object has Class "fd"
}
\description{
Check that an argument is a functional data object.
}
\usage{
is.fd(fdobj)
}
\arguments{
\item{fdobj}{
an object to be checked.
}
}
\value{
a logical value:
\code{TRUE} if the class is correct, \code{FALSE} otherwise.
}
\seealso{
is.basis, is.fdPar, is.Lfd
}
% docclass is function
\keyword{smooth}
