\name{is.fdPar}
\alias{is.fdPar}
\title{
  Confirm Object has Class "fdPar"
}
\description{
Check that an argument is a functional parameter object.
}
\usage{
is.fdPar(fdParobj)
}
\arguments{
\item{fdParobj}{
an object to be checked.
}
}
\value{
a logical value:
\code{TRUE} if the class is correct, \code{FALSE} otherwise.
}
\seealso{
\code{\link{is.basis}}, 
\code{\link{is.fd}}, 
\code{\link{is.Lfd}}
}
% docclass is function
\keyword{smooth}
