\name{is.fdSmooth}
\alias{is.fdSmooth}
\title{
  Confirm Object has Class "fdSmooth"
}
\description{
Check that an argument is a functional parameter object.
}
\usage{
is.fdSmooth(fdSmoothobj)
}
\arguments{
\item{fdSmoothobj}{
an object to be checked.
}
}
\value{
a logical value:
\code{TRUE} if the class is correct, \code{FALSE} otherwise.
}
\seealso{
\code{\link{is.basis}},
\code{\link{is.fd}},
\code{\link{is.Lfd}},
\code{\link{is.fdPar}}
}
% docclass is function
\keyword{smooth}