\name{linesFd}
\alias{linesFd}
\title{Add Lines from Functional Data to a Plot }
\description{
Lines defined by functional observations are added to an existing plot. 
}
\usage{
linesFd(fd, Lfd=0, ...)
}
\arguments{
  \item{fd }{A univariate functional data object. }
  \item{Lfd }{Either a nonnegative integer or a linear differential operator object. If present, the derivative or the value of applying the operator is evaluated rather than the functions themselves. }
  \item{... }{Additional arguments such as axis titles and so forth that can be used in all plotting programs. }
}
\details{
}
\value{
Lines added to an existing plot. 
}
\references{}
\note{}
\author{}
\seealso{
plotFd, plotfit.fd 
}
\examples{
}
\keyword{}
