\name{meanFd}
\alias{meanFd}
\title{Mean of Functional Data }
\description{
Evaluate the mean of a set of functions in a functional data object. 
}
\usage{
meanFd(fd)
}
\arguments{
  \item{fd }{A functional data object. }
}
\details{
}
\value{
Returns a functional data object with a single replication that contains the mean of the functions in the object fd. 
}
\references{}
\note{}
\author{}
\seealso{
std.fd, sumFd, center.fd }
\examples{}
\keyword{}
