\name{nondurables}
\alias{nondurables}
\non_function{}
\title{
  Nondurable goods index
}
\description{
  US nondurable goods index time series, January 1919 to January 2000.  
}
\format{
    An object of class 'ts'.  
}
\details{

}
\source{
  Ramsay, James O., and Silverman, Bernard W. (2006), \emph{Functional
    Data Analysis, 2nd ed.}, Springer, New York. 

  Ramsay, James O., and Silverman, Bernard W. (2002), \emph{Applied
    Functional Data Analysis}, Springer, New York, ch. 3. 
}
%\seealso{\code{\link{eval.penalty}}}
\examples{
plot(nondurables, log="y")
}
% docclass is function
\keyword{datasets}
