\name{printFd}
\alias{printFd}
\title{Print a Functional Data Object }
\description{
Prints a functional data object. 
}
\usage{
printFd(fd, ...)
}
\arguments{
  \item{fd }{A functional data object. }
  \item{... }{The additional optional arguments that are permitted for all print functions. }
}
\details{
}
\value{
The dimensions of the coefficient array are displayed, as well as the coefficient array itself. Then the characteristics of the basis are printed. 
}
\references{}
\note{}
\author{}
\seealso{}
\examples{
}
\keyword{}
