\name{std.fd}
\alias{std.fd}
\title{Standard Deviation of Functional Data }
\description{
Evaluate the standard deviation of a set of functions in a functional data object. 
}
\usage{
std.fd(fd)
}
\arguments{
  \item{fd }{A functional data object. }
}
\details{
}
\value{
Returns a functional data object with a single replication that contains the standard deviation of the one or several functions in the object fd. 

}
\references{}
\note{}
\author{}
\seealso{
meanFd, sumFd, center.fd 
}
\examples{}
\keyword{}
