\name{sumFd}
\alias{sumFd}
\title{Sum of Functional Data }
\description{
Evaluate the sum of a set of functions in a functional data object. 
}
\usage{
sumFd(fd)
}
\arguments{
  \item{fd }{A functional data object. }
}
\details{
}
\value{
Returns a functional data object with a single replication that contains the sum of the functions in the object fd. 
}
\references{}
\note{}
\author{}
\seealso{
std.fd, meanFd, center.fd 
}
\examples{}
\keyword{}
