\name{summary.basisfd}
\alias{summary.basisfd}
\title{
  Summarize a Functional Data Object
}
\description{
Provide a compact summary of the characteristics of a
functional data object.
}
\usage{
summary.basisfd(object, ...)
}
\arguments{
\item{object}{
a functional data object.
}
\item{...}{Other arguments to match generic}
}
\value{
a displayed summary of the bivariate functional data object.
}
% docclass is function
\keyword{smooth}
