\name{summary.bifd}
\alias{summary.bifd}
\title{
  Summarize a Bivariate Functional Data Object
}
\description{
Provide a compact summary of the characteristics of a
bivariate functional data object.
}
\usage{
summary.bifd(object, ...)
}
\arguments{
\item{object}{
a bivariate functional data object.
}
\item{...}{Other arguments to match generic}
}
\value{
a displayed summary of the bivariate functional data object.
}
% docclass is function
\keyword{smooth}
