\name{summary.fd}
\alias{summary.fd}
\title{
  Summarize a Functional Data Object
}
\description{
Provide a compact summary of the characteristics of a functional
data object.
}
\usage{
summary.fd(object,...)
}
\arguments{
\item{object}{
a functional data object.
}
\item{...}{Other arguments to match generic}
}
\value{
a displayed summary of the functional data object.
}
% docclass is function
\keyword{smooth}
