\name{summary.fdPar}
\alias{summary.fdPar}
\title{
  Summarize a Functional Parameter Object
}
\description{
Provide a compact summary of the characteristics of a functional
parameter object.
}
\usage{
summary.fdPar(object, ...)
}
\arguments{
\item{object}{
a functional parameter object.
}
\item{...}{Other arguments to match generic}
}
\value{
a displayed summary of the functional parameter object.
}
% docclass is function
\keyword{smooth}
