\name{summaryFd}
\alias{summaryFd}
\title{Summarize a Functional Data Object }
\description{
Provided a compact summary of the characteristics of a funtional data object. 
}
\usage{
summaryFd(fd, ...)
}
\arguments{
  \item{fd }{A functional data object. }
  \item{... }{The additional arguments that may be supplied to the usual summary function. }
}
\details{
}
\value{
A displayed summary of the functional data object. 
}
\references{}
\note{}
\author{}
\seealso{}
\examples{
}
\keyword{}
