\name{varmx.pca.fd}
\alias{varmx.pca.fd}
\title{Rotation of Functional Principal Components with VARIMAX Criterion }
\description{
Principal components are often easier to interpret if they are rotated. Among the many possible ways in which this rotation can be defined, the VARIMAX criterion seems to give satisfactory results most of the time. 
}
\usage{
varmx.pca.fd(pcafd, nharm=scoresd[2], nx=101)
}
\arguments{
  \item{pcafd }{A principal components analysis object of class pca.fd. }
  \item{nharm }{The number of principal components or harmonics to be rotated. }
  \item{nx }{The number of argument values used to define the harmonics to be rotated. }
}
\details{
}
\value{
A princpal components analysis object of class pca.fd. 
}
\references{}
\note{}
\author{}
\seealso{}
\examples{
}
\keyword{}
