\name{arithmetic.basisfd}
\alias{arithmetic.basisfd}
\alias{==.basisfd}
\title{
  Arithmatic on functional basis objects
}
\description{
  Arithmatic on functional basis objects
}
\usage{
\method{==}{basisfd}(basis1, basis2)
}
\arguments{
  \item{basis1, basis2}{
    functional basis object
  }
}

\value{
  basisobj1 == basisobj2 returns a logical scalar.
}
\references{
  Ramsay, James O., and Silverman, Bernard W. (2006), \emph{Functional
    Data Analysis, 2nd ed.}, Springer, New York.

  Ramsay, James O., and Silverman, Bernard W. (2002), \emph{Applied
    Functional Data Analysis}, Springer, New York.
}
\seealso{
\code{\link{basisfd}},
\code{\link{basisfd.product}}
\code{\link{arithmetic.fd}}
}
% docclass is function
\keyword{smooth}

