\name{create.constant.basis}
\alias{create.constant.basis}
\title{Create a Constant Basis}
\description{
Create a constant basis object, defining a single basis function whose value is everywhere 1.0. 
}
\usage{
create.constant.basis(rangeval=c(0, 1))
}
\arguments{
  \item{rangeval}{A vector of length 2 containing the initial and final values of argument t defining the interval over which the functional data object can be evaluated. However, this is seldom used since the value of the basis function does not depend on the range or any argument values. }
  \item{}{}
  \item{}{}
}
\details{
}
\value{
A basis object of the type "const". 
}
\references{}
\note{}
\author{}
\seealso{}
\examples{}
\keyword{}
