\name{getbasisrange}
\alias{getbasisrange}
\title{
  Extract the range from a basis object 
}
\description{
  Extracts the 'range' component from basis object 'basisobj'.  
}
\usage{
getbasisrange(basisobj) 
}
\arguments{
  \item{basisobj}{
    a functional basis object 
  }
}

\value{
  a numeric vector of length 2 
}
%\seealso{}
%\examples{}
% docclass is function
\keyword{attribute}
