\name{is.basis}
\alias{is.basis}
\title{
  Confirm Object is Class "Basisfd"
}
\description{
Check that an argument is a basis object.
}
\usage{
is.basis(basisobj)
}
\arguments{
\item{basisobj}{
an object to be checked.
}
}
\value{
a logical value:
\code{TRUE} if the class is correct, \code{FALSE} otherwise.
}
\seealso{
\code{\link{is.fd}}, 
\code{\link{is.fdPar}}, 
\code{\link{is.Lfd}}
}
% docclass is function
\keyword{smooth}
