\name{is.Lfd}
\alias{is.Lfd}
\title{
  Confirm Object has Class "Lfd"
}
\description{
Check that an argument is a linear differential operator object.
}
\usage{
is.Lfd(Lfdobj)
}
\arguments{
\item{Lfdobj}{
an object to be checked.
}
}
\value{
a logical value:
\code{TRUE} if the class is correct, \code{FALSE} otherwise.
}
\seealso{
\code{\link{is.basis}}, 
\code{\link{is.fd}}, 
\code{\link{is.fdPar}}
}
% docclass is function
\keyword{smooth}
