\name{lip}
\alias{lip}
\non_function{}
\title{
  Lip motion 
}
\description{
  a matrix of dimensions (51, 20) giving ... ??? 
}
\format{
}
\details{
}
\source{
  Ramsay, James O., and Silverman, Bernard W. (2002), \emph{Applied
    Functional Data Analysis}, Springer, New York, ch. 10. 
}
%\seealso{\code{\link{eval.penalty}}}
\examples{

}
% docclass is function
\keyword{datasets}
