\name{summary.basisfd}
\alias{summary.basisfd}
\title{
  Summarize a Functional Data Object
}
\description{
Provide a compact summary of the characteristics of a
functional data object.
}
\usage{
\method{summary}{basisfd}(object, ...)
}
\arguments{
\item{object}{
a functional data object (i.e., of class 'basisfd').
}
\item{...}{Other arguments to match generic}
}
\value{
a displayed summary of the bivariate functional data object.
}
\references{
  Ramsay, James O., Hooker, Giles, and Graves, Spencer (2009),
    \emph{Functional data analysis with R and Matlab}, Springer, New York.
  
  Ramsay, James O., and Silverman, Bernard W. (2005), 
    \emph{Functional Data Analysis, 2nd ed.}, Springer, New York.
  
  Ramsay, James O., and Silverman, Bernard W. (2002), 
    \emph{Applied Functional Data Analysis}, Springer, New York.
}
\keyword{smooth}
