\name{summary.bifd}
\alias{summary.bifd}
\title{
  Summarize a Bivariate Functional Data Object
}
\description{
Provide a compact summary of the characteristics of a
bivariate functional data object.
}
\usage{
\method{summary}{bifd}(object, ...)
%summary.bifd(object, ...)
}
\arguments{
\item{object}{
a bivariate functional data object.
}
\item{...}{Other arguments to match the generic function for 'summary'}
}
\value{
a displayed summary of the bivariate functional data object.
}
\references{
  Ramsay, James O., Hooker, Giles, and Graves, Spencer (2009),
    \emph{Functional data analysis with R and Matlab}, Springer, New York.
  
  Ramsay, James O., and Silverman, Bernard W. (2005), 
    \emph{Functional Data Analysis, 2nd ed.}, Springer, New York.
  
  Ramsay, James O., and Silverman, Bernard W. (2002), 
    \emph{Applied Functional Data Analysis}, Springer, New York.
}
\seealso{
  \code{\link{summary}},
}
% docclass is function
\keyword{smooth}
