\name{summary.bifd}
\alias{summary.bifd}
\title{
  Summarize a Bivariate Functional Data Object
}
\description{
Provide a compact summary of the characteristics of a
bivariate functional data object.
}
\usage{
\method{summary}{bifd}(object, ...)
%summary.bifd(object, ...)
}
\arguments{
\item{object}{
a bivariate functional data object.
}
\item{...}{Other arguments to match the generic function for 'summary'}
}
\value{
a displayed summary of the bivariate functional data object.
}
\seealso{
\code{\link{summary}},
}
% docclass is function
\keyword{smooth}
