\name{summary.fdPar}
\alias{summary.fdPar}
\title{
  Summarize a Functional Parameter Object
}
\description{
Provide a compact summary of the characteristics of a functional
parameter object (i.e., class 'fdPar').
}
\usage{
\method{summary}{fdPar}(object, ...)
%summary.fdPar(object, ...)
}
\arguments{
\item{object}{
a functional parameter object.
}
\item{...}{Other arguments to match the generic 'summary' function}
}
\value{
a displayed summary of the functional parameter object.
}
\references{
  Ramsay, James O., Hooker, Giles, and Graves, Spencer (2009),
    \emph{Functional data analysis with R and Matlab}, Springer, New York.
  
  Ramsay, James O., and Silverman, Bernard W. (2005), 
    \emph{Functional Data Analysis, 2nd ed.}, Springer, New York.
  
  Ramsay, James O., and Silverman, Bernard W. (2002), 
    \emph{Applied Functional Data Analysis}, Springer, New York.
}
\seealso{
  \code{\link{summary}},
}
\keyword{smooth}
