\name{summary.fdPar}
\alias{summary.fdPar}
\title{
  Summarize a Functional Parameter Object
}
\description{
Provide a compact summary of the characteristics of a functional
parameter object (i.e., class 'fdPar').
}
\usage{
\method{summary}{fdPar}(object, ...)
%summary.fdPar(object, ...)
}
\arguments{
\item{object}{
a functional parameter object.
}
\item{...}{Other arguments to match the generic 'summary' function}
}
\value{
a displayed summary of the functional parameter object.
}
\seealso{
\code{\link{summary}},
}
% docclass is function
\keyword{smooth}
