\name{summary.Lfd}
\alias{summary.Lfd}
\title{
  Summarize a Linear Differential Operator Object
}
\description{
Provide a compact summary of the characteristics of a
linear differential operator object.
}
\usage{
\method{summary}{Lfd}(object, ...)
%summary.Lfd(object, ...)
}
\arguments{
\item{object}{
a linear differential operator object.
}
\item{...}{Other arguments to match the generic 'summary' function}
}
\value{
a displayed summary of the linear differential operator object.
}
\references{
  Ramsay, James O., Hooker, Giles, and Graves, Spencer (2009),
    \emph{Functional data analysis with R and Matlab}, Springer, New York.
  
  Ramsay, James O., and Silverman, Bernard W. (2005), 
    \emph{Functional Data Analysis, 2nd ed.}, Springer, New York.
  
  Ramsay, James O., and Silverman, Bernard W. (2002), 
    \emph{Applied Functional Data Analysis}, Springer, New York.
}
\seealso{
  \code{\link{summary}},
}
% docclass is function
\keyword{smooth}
