\name{summary.Lfd}
\alias{summary.Lfd}
\title{
  Summarize a Linear Differential Operator Object
}
\description{
Provide a compact summary of the characteristics of a
linear differential operator object.
}
\usage{
\method{summary}{Lfd}(object, ...)
%summary.Lfd(object, ...)
}
\arguments{
\item{object}{
a linear differential operator object.
}
\item{...}{Other arguments to match the generic 'summary' function}
}
\value{
a displayed summary of the linear differential operator object.
}
\seealso{
\code{\link{summary}},
}
% docclass is function
\keyword{smooth}
